pyproject.toml
[tool.poetry]
name = "murphy threads backend"
version = "0.1.0"
description = "Backend Of Murphy Threads Website"
authors = ["Aditya Bhushan <adityapanto123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.3"
djangorestframework = "^3.14.0"
markdown = "^3.4.3"
django-filter = "^23.2"
psycopg2-binary = "^2.9.6"
stripe = "^5.5.0"
instamojo-wrapper = "^1.2.0"
paypalrestsdk = "^1.13.1"
djangorestframework-simplejwt = "^5.2.2"
django-cors-headers = "^4.2.0"
django-dotenv = "^1.4.2"
python-dotenv = "^1.0.1"
ipinfo-django = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
